pull out utility for dealing with help input and force yargs to use console.log for logging help
