+++ packages/core/lib/console-child.js
@@ -4,14 +4,23 @@ const yargs = require("yargs");
 const path = require("path");
 const {
   deriveConfigEnvironment,
-  parseQuotesAndEscapes
+  parseQuotesAndEscapes,
+  displayGeneralHelp
 } = require("./command-utils");
+const { handleHelpInput } = require("./cliHelp");
 
 // we split off the part Truffle cares about and need to convert to an array
 const input = process.argv[2].split(" -- ");
 const escapeCharacters = path.sep === "\\" ? "^`" : "\\"; //set escape character
 //based on current OS; backslash for Unix, caret or grave for Windows
 const inputStrings = parseQuotesAndEscapes(input[1], escapeCharacters); //note this shouldn't error since it's a recomputation
+
+// handle cases where input indicates the user wants to access Truffle's help
+const { displayHelp } = handleHelpInput({ inputStrings });
+if (displayHelp) {
+  displayGeneralHelp();
+  process.exit();
+}
 
 // we need to make sure this function exists so ensjs doesn't complain as it requires
 // getRandomValues for some functionalities - webpack strips out the crypto lib
